Remove all reference to MongoDB and remove jskos-server
The application makes use of jskos-server (more specific: a single
Changes API endpoint) for updates only. Either there such API endpoint
or bartoc-search should works but without live updates. The internals of
jskos-server such as any reference to MongoDB should not be concern of
bartoc-search.

In contrast Solr and Redis are required by bartoc-search but not
connected to any other component, so their setup should be helped with.

This commit tries to remove all reference to MongoDB and jskos-server
except documentation how to configure the Changes API endpoint. It also
adds some documentation how to get started in setting up BARTOC search
with local Solr and Redis. This may still be simplified by using
`localhost` as default host in `config/config.default.json`.